+++ src/routers/usersRouter.ts
@@ -4,10 +4,12 @@ import { usersHandlers } from '../handlers/usersHandlers';
 import {
 	requireAuthentication,
 	requireChangemakerPermission,
+	requireDataProviderPermission,
 	requireFunderPermission,
 } from '../middleware';
 import { Permission } from '../types';
 import { userFunderPermissionsHandlers } from '../handlers/userFunderPermissionsHandlers';
+import { userDataProviderPermissionsHandlers } from '../handlers/userDataProviderPermissionsHandlers';
 
 const usersRouter = express.Router();
 
@@ -21,6 +23,11 @@ usersRouter.delete(
 	'/:userKeycloakUserId/changemakers/:changemakerId/permissions/:permission',
 	requireChangemakerPermission(Permission.MANAGE),
 	userChangemakerPermissionsHandlers.deleteUserChangemakerPermission,
+);
+usersRouter.put(
+	'/:userKeycloakUserId/dataProviders/:dataProviderShortCode/permissions/:permission',
+	requireDataProviderPermission(Permission.MANAGE),
+	userDataProviderPermissionsHandlers.putUserDataProviderPermission,
 );
 usersRouter.put(
 	'/:userKeycloakUserId/funders/:funderShortCode/permissions/:permission',